Move branding.dart into config/
