+++ lib/features/audio_player/presentation/widgets/miniplayer.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:vidya_music/core/utils/branding.dart';
+import 'package:vidya_music/config/branding.dart';
 import 'package:vidya_music/core/utils/extensions/build_context_l10n_ext.dart';
 import 'package:vidya_music/features/audio_player/presentation/widgets/player.dart';
 import 'package:vidya_music/features/audio_player/presentation/widgets/player_controls.dart';
