+++ lib/main.dart
@@ -6,8 +6,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:just_audio_background/just_audio_background.dart';
 import 'package:provider/provider.dart';
+import 'package:vidya_music/config/branding.dart';
 import 'package:vidya_music/core/theme/app_theme.dart';
-import 'package:vidya_music/core/utils/branding.dart';
 import 'package:vidya_music/core/utils/utils.dart';
 import 'package:vidya_music/features/audio_player/data/datasources/audio_player_service.dart';
 import 'package:vidya_music/features/audio_player/presentation/bloc/audio_player_cubit.dart';
